Make some more refactor for createElement function to narrow the returned types
