+++ week-1-ts-ecosystem/w01e02-enter-the-world-of-typescript/src/model/index.ts
@@ -1,0 +1,14 @@
+export type Currency = 'PLN' | 'USD'
+export type Unit = 'kg' | 'g'
+
+export interface Price {
+  value: number;
+  currency: Currency;
+}
+
+export interface Item {
+  name: string;
+  amount: number;
+  unit: Unit;
+  price: Price;
+}
